Move human controller to a new script
